fastapi_migration/backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
alembic>=1.11.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
reportlab>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
pg8000>=1.29.0
asyncpg>=0.27.0
psycopg2-binary>=2.9.6
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
httpx>=0.24.0
markdown>=3.4.3
pytest>=7.3.1
pytest-asyncio>=0.21.0
PyJWT>=2.6.0
pydantic_settings

# Google Cloud SQL dependencies
cloud-sql-python-connector[pg8000]>=1.2.0
google-cloud-secret-manager>=2.13.0